Moving files around and implementing more of the client manager.
