+++ Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 
 [dependencies]
 regex = "1"
+uuid = "0.8"
 crossbeam = "0.7"
 crossbeam-channel = "0.4"
 crossbeam-utils = "0.7"
@@ -19,11 +20,8 @@ zeroize = "1.1.0"
 crossterm = "0.17.7"
 clap = "3.0.0-beta.1"
 log = "0.4"
-cursive = { version = "0.15.0", default-features = false, features = ["crossterm-backend"]}
-openssl = { version = "0.10", features = ["vendored"] }
-rustls = "0.18.1"
-webpki = "0.21.3"
-webpki-roots = "0.20.0"
+serde = { version = "1.0", features = ["derive"] }
+url = "2.2.0"
 
 
 [profile.dev]
