+++ src/app/bundle.rs
@@ -1,0 +1,11 @@
+
+/**
+ * Bundle: inspired from NSBundle on macOS
+ */
+struct Bundle {
+  location:
+}
+
+impl Bundle {
+  
+}
