+++ src/lib/server/config.rs
@@ -1,0 +1,9 @@
+
+pub struct ServerConfig {
+  pub name: String,
+  pub address: String,
+  pub owner: String,
+
+  pub host: String,
+  pub port: u16,
+}
