+++ src/lib/server/mod.rs
@@ -1,0 +1,5 @@
+pub mod client_management;
+pub mod server;
+pub mod server_v3;
+
+pub struct Server {}
