--- src/server/client/mod.rs
@@ -1,2 +1,0 @@
-pub mod client_profile;
-pub mod client_v3;
