--- src/server/mod.rs
@@ -1,4 +1,0 @@
-pub mod server_v3;
-pub mod client;
-pub mod server_profile;
-pub mod ui;
