--- src/server/ui/mod.rs
@@ -1,4 +1,0 @@
-pub mod server_view_controller;
-pub mod control_panel;
-pub mod about_panel;
-pub mod main_menu;
